API and a dumb implementation of stopwatches.
A stopwatch is an accumulating timer that keeps track of time durations in
an efficient way. The current implementation is a trivial prototype and uses
converted uptime timestamps, but the future plan is to use raw values from
timecounters for cheapest possible timestamping. The contents of struct
stopwatch are opaque to anyone other than subr_stopwatch.c, they are only
exposed because it's the only efficient way to embed them in other structs.

The API is:
 - stopwatch_reset()
    Resets a stopwatch to zero elapsed time.
 - stopwatch_start()
    Start accumulating time in the stopwatch (no reset).
 - stopwatch_stop()
    Stop accumulating time in the stopwatch.
 - stopwatch_handover(from, to)
    Equivalent to "topwatch_stop(from); stopwatch_start(to);" using the
    same timestamp for the start and stop times.
 - stopwatch_snapshot(from, to)
    Equivalent to copying the contents of "from" into "to" and calling
    "stopwatch_stop(to);". Notice that it's up to the caller to make sure
    that the stopwatch "from" is running. Calling stopwatch_snapshot on a
    non-running stopwatch will return garbage.
 - stopwatch_to_ts(),
   stopwatch_to_tv(),
   stopwatch_to_s()
    Convert the internal value of a stopped stopwatch into a timespec,
    timeval and seconds (respectively). Calling any of these functions
    on a currently running stopwatch will return garbage.

(To not stretch the analogy too much I resisted the urge of calling it
 stopwatch_lap instead of stopwatch_snapshot.)