Convert p_rtime into a stopwatch with a sledgehammer.
We stop keeping track of spc_runtime since the per-proc p_runtime
has all the data we need. Be careful about starting and stopping
the stopwatch in the right places (the old code could get away
with being sloppy).

What's making it difficult to do it nicely is struct tusage and how
it's being kept track of. This will probably be redesigned next.