Integrate the userland part of FreeBSD sysctl with a sledgehammer.
We let it do part of the command line parsing and the "all" listing
and if that doesn't know what to do, we retry using the old code instead.

Right now this breaks "sysctl kern", so I guess the requirement should
be to either convert a whole tree, or nothing. Or something. Good enough
for now.